[mutable] Fixes in mesh clipping for meshes whose buffers have not been cache optimized. #rnx [REVIEW] [at]genis.sole #rb genis.sole
[CL 29176574 by jordi rovira in ue5-main branch]

+++ Engine/Plugins/Experimental/Mutable/Source/MutableRuntime/Private/MuR/OpMeshApplyLayout.h
@@ -195,7 +195,7 @@ namespace mu
 			}
 	
 			// Remove
-			MeshRemoveVerticesWithMap( pApplied, VertexMask.GetData(), VertexMask.Num());
+			MeshRemoveVerticesWithMap( pApplied, VertexMask.GetData(), (uint32)VertexMask.Num());
 		}
 
 		//
